+++ modules/calcite/src/main/java/org/apache/ignite/internal/processors/query/calcite/util/IgniteMethod.java
@@ -101,7 +101,7 @@ public enum IgniteMethod {
     /** See {@link IgniteSqlFunctions#skipFirstArgument(Object, Object)}. **/
     SKIP_FIRST_ARGUMENT(IgniteSqlFunctions.class, "skipFirstArgument", Object.class, Object.class),
 
-    /** See {@link IgniteSqlFunctions#bitwise(SqlKind, Number, Number)}. **/
+    /** See {@link IgniteMath#bitwise(SqlKind, Number, Number)}. **/
     BITWISE(IgniteSqlFunctions.class, "bitwise", SqlKind.class, Number.class, Number.class);
 
     /** */
